rest-model-based WIP
