--- ataskq/db_handler/__init__.py
@@ -1,1 +1,0 @@
-from .db_handler import DBHandler, EAction
