+++ ataskq/test_models.py
@@ -1,8 +1,7 @@
 import pytest
 
 from .models import Model, __MODELS__, Job
-from .handler import Handler, from_connection_str
-from .ihandler import register_handler, unregister_handler
+from .ihandler import Handler, from_connection_str, register_handler, unregister_handler
 
 
 @pytest.fixture(scope="function")
