+++ contrib/run_test.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import context
 from ataskq.test_ataskq import test_state_kwargs_job_delete_cascade as t
-from ataskq.db_handler.db_handler import set_connection_log
+from ataskq.ihandler.db_handler import set_connection_log
 
 if __name__ == "__main__":
     db_path = "./sqlite.local.db"
